test(tool-offsets): one toolchanger fixture, not one per file
The two panel test files had grown byte-identical setup: the mock's
toolchanger persona, the discovery the capability question reads, ToolState's
four tools and a real MoonrakerAPI over the mock. Only the base class differed
— one drives the run with no widgets, the other builds the XML — so the base
is now the template parameter and the setup lives once, as busy_guard_fixture.h
does for its three sharers. The mock's own test keeps its fixture: it tests the
simulator and wants none of this around it.

mutation: with the shared fixture in place, all five production mutations still
go red (the invalidate in abort, the nothing-dirty return, the idle-edge
re-read, the ui_alive_ guard, and send_save() as a bare return).
